+++ dtslint/ts4.1/Semigroup.ts
@@ -1,4 +1,5 @@
 import * as _ from '../../src/Semigroup'
+import { pipe } from '../../src/function'
 
 //
 // getTupleSemigroup
@@ -10,5 +11,4 @@ _.getTupleSemigroup(_.semigroupString, _.semigroupSum, _.semigroupAll) // $Expec
 // fold
 //
 
-_.fold(_.semigroupString)('', ['a']) // $ExpectType string
-_.fold(_.semigroupString)('') // $ExpectType (as: readonly string[]) => string
+pipe(['a'], _.fold(_.semigroupString)('')) // $ExpectType string